pyproject.toml
[tool.poetry]
name = "esg-summary-generator"
version = "0.1.0"
description = ""
authors = ["A-baoYang <rubyyang0615@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.6"
openai = "^1.12.0"
langchain-openai = "^0.0.5"
python-dotenv = {extras = ["cli"], version = "^1.0.1"}
gradio = "^4.18.0"
rag-sdk = {git = "https://github.com/ESGTrans/rag_sdk.git", rev = "main"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
